+++ packages/draggable/types.ts
@@ -1,4 +1,5 @@
 import {Avatar} from './avatar';
+import {Point} from './utils/point';
 
 export type Axis = 'both' | 'horizontal' | 'vertical';
 
@@ -12,7 +13,11 @@ export interface DraggableAttachOpts<T> {
     draggingClass?: string;
     draggingClassBody?: string;
     minDragStartDistance?: number;
+    touchAction?: string | null;
+    customScroll?: CustomScroll
 }
+
+export type CustomScroll = ((startPosition: Point, currentPosition: Point) => void) | boolean;
 
 export interface DraggableInterface {
     _id: number;