+++ packages/dropzone/types.ts
@@ -2,6 +2,7 @@ import {Acceptor} from './acceptor';
 
 export interface DropzoneAttachOpts {
     acceptor?: Acceptor;
+    exact?: boolean;
     overClass?: string;
     invalidClass?: string;
 }